Separating Http from Rpc and Rest
