+++ HttpServer/HttpServer/HttpControllerAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Neo.Plugins.HttpServer
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class HttpControllerAttribute : Attribute
+    {
+        public string Path { get; set; }
+    }
+}
