+++ HttpServer/HttpServer/HttpMethodAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Neo.Plugins.HttpServer
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public class HttpMethodAttribute : Attribute
+    {
+        public string Path { get; set; }
+    }
+}
