--- RpcServer/RpcServer/RpcControllerAttribute.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Neo.Plugins.RpcServer
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public class RpcControllerAttribute : Attribute
-    {
-        public string Name { get; set; }
-    }
-}
