--- RpcServer/RpcServer/RpcTargetAndMethod.cs
@@ -1,11 +1,0 @@
-﻿using System.Reflection;
-
-namespace Neo.Plugins.RpcServer
-{
-    internal class RpcTargetAndMethod
-    {
-        public object Target { get; set; }
-
-        public MethodInfo Method { get; set; }
-    }
-}
